// spark-event-ui/src/app/rest/model/sparkAppSummaryDTO.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SparkEnvironmentUpdateEventSummaryDTO } from './sparkEnvironmentUpdateEventSummaryDTO';
import { SparkAppMetricsDTO } from './sparkAppMetricsDTO';
import { SparkApplicationStartEventSummaryDTO } from './sparkApplicationStartEventSummaryDTO';
import { SparkTopLevelJobExecEventSummaryDTO } from './sparkTopLevelJobExecEventSummaryDTO';
import { TemplateDictionariesEntriesDTO } from './templateDictionariesEntriesDTO';
import { SparkApplicationEndEventSummaryDTO } from './sparkApplicationEndEventSummaryDTO';
import { SparkResourceProfileAddedEventSummaryDTO } from './sparkResourceProfileAddedEventSummaryDTO';
import { RepeatedLongestTopLevelJobExecsPerCallSiteDTO } from './repeatedLongestTopLevelJobExecsPerCallSiteDTO';
import { SparkSQLExecutionEventSummaryDTO } from './sparkSQLExecutionEventSummaryDTO';
import { SparkApplicationConfigDTO } from './sparkApplicationConfigDTO';
import { RepeatedLongestSqlExecsPerCallSiteDTO } from './repeatedLongestSqlExecsPerCallSiteDTO';
import { SparkLogStartEventSummaryDTO } from './sparkLogStartEventSummaryDTO';


export interface SparkAppSummaryDTO { 
    clusterName?: string;
    sparkAppName?: string;
    sparkAppConfig?: SparkApplicationConfigDTO;
    appMetrics?: SparkAppMetricsDTO;
    applicationStart?: SparkApplicationStartEventSummaryDTO;
    applicationEnd?: SparkApplicationEndEventSummaryDTO;
    logStart?: SparkLogStartEventSummaryDTO;
    resourceProfileAdded?: SparkResourceProfileAddedEventSummaryDTO;
    envUpdate?: SparkEnvironmentUpdateEventSummaryDTO;
    longestSqlExecs?: Array<SparkSQLExecutionEventSummaryDTO>;
    longestTopLevelJobExecs?: Array<SparkTopLevelJobExecEventSummaryDTO>;
    repeatedLongestSqlExecsPerCallSites?: Array<RepeatedLongestSqlExecsPerCallSiteDTO>;
    repeatedLongestTopLevelJobExecsPerCallSites?: Array<RepeatedLongestTopLevelJobExecsPerCallSiteDTO>;
    templateDictionariesEntries?: TemplateDictionariesEntriesDTO;
}