// spark-event-ui/src/app/rest/model/sqlExecutionEventsDTO.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { FindByNum200Response } from './findByNum200Response';


export interface SqlExecutionEventsDTO { 
    sqlId?: number;
    startAppEvents?: Array<FindByNum200Response>;
    startSqlEventNum?: number;
    events?: Array<FindByNum200Response>;
    endSqlEventNum?: number;
}